requirements.txt
appdirs==1.4.3
Django==1.8
djangorestframework==3.6.2
mysql-python==1.2.5
packaging==16.8
pyparsing==2.2.0
six==1.10.0
